refactor: only fetch necessary data in notifications context
