+++ apps/hash-frontend/src/shared/draft-entities-context.tsx
@@ -1,4 +1,4 @@
-import { useMutation, useQuery } from "@apollo/client";
+import { useQuery } from "@apollo/client";
 import {
   currentTimeInstantTemporalAxes,
   mapGqlSubgraphFieldsFragmentToSubgraph,
@@ -10,34 +10,22 @@ import {
   createContext,
   FunctionComponent,
   PropsWithChildren,
-  useCallback,
   useContext,
   useMemo,
   useState,
 } from "react";
 
 import {
-  ArchiveEntityMutation,
-  ArchiveEntityMutationVariables,
   StructuralQueryEntitiesQuery,
   StructuralQueryEntitiesQueryVariables,
-  UpdateEntityMutation,
-  UpdateEntityMutationVariables,
 } from "../graphql/api-types.gen";
-import {
-  archiveEntityMutation,
-  structuralQueryEntitiesQuery,
-  updateEntityMutation,
-} from "../graphql/queries/knowledge/entity.queries";
-import { useNotifications } from "./notifications-context";
+import { structuralQueryEntitiesQuery } from "../graphql/queries/knowledge/entity.queries";
 
 export type DraftEntitiesContextValue = {
   draftEntities?: Entity[];
   draftEntitiesSubgraph?: Subgraph<EntityRootType>;
   loading: boolean;
   refetch: () => Promise<void>;
-  discardDraftEntity: (params: { draftEntity: Entity }) => Promise<void>;
-  acceptDraftEntity: (params: { draftEntity: Entity }) => Promise<Entity>;
 };
 
 export const DraftEntitiesContext =
@@ -110,100 +98,6 @@ export const DraftEntitiesContextProvider: FunctionComponent<
     [draftEntitiesSubgraph],
   );
 
-  const { notifications, archiveNotification, markNotificationAsRead } =
-    useNotifications();
-
-  const archiveRelatedNotifications = useCallback(
-    async (params: { draftEntity: Entity }) => {
-      const relatedNotifications = notifications?.filter(
-        (notification) =>
-          notification.occurredInEntity.metadata.recordId.entityId ===
-          params.draftEntity.metadata.recordId.entityId,
-      );
-
-      if (!relatedNotifications) {
-        return;
-      }
-
-      await Promise.all(
-        relatedNotifications.map((notification) => {
-          return archiveNotification({
-            notificationEntityId:
-              notification.entity.metadata.recordId.entityId,
-          });
-        }),
-      );
-    },
-    [notifications, archiveNotification],
-  );
-
-  const [archiveEntity] = useMutation<
-    ArchiveEntityMutation,
-    ArchiveEntityMutationVariables
-  >(archiveEntityMutation);
-
-  const discardDraftEntity = useCallback(
-    async (params: { draftEntity: Entity }) => {
-      await archiveRelatedNotifications(params);
-
-      await archiveEntity({
-        variables: {
-          entityId: params.draftEntity.metadata.recordId.entityId,
-        },
-      });
-
-      await refetch();
-    },
-    [archiveEntity, archiveRelatedNotifications, refetch],
-  );
-
-  const [updateEntity] = useMutation<
-    UpdateEntityMutation,
-    UpdateEntityMutationVariables
-  >(updateEntityMutation);
-
-  const markRelatedGraphChangeNotificationsAsRead = useCallback(
-    async (params: { draftEntity: Entity }) => {
-      const relatedGraphChangeNotifications =
-        notifications?.filter(
-          ({ kind, occurredInEntity }) =>
-            kind === "graph-change" &&
-            occurredInEntity.metadata.recordId.entityId ===
-              params.draftEntity.metadata.recordId.entityId,
-        ) ?? [];
-
-      await Promise.all(
-        relatedGraphChangeNotifications.map((notification) =>
-          markNotificationAsRead({ notification }),
-        ),
-      );
-    },
-    [notifications, markNotificationAsRead],
-  );
-
-  const acceptDraftEntity = useCallback(
-    async (params: { draftEntity: Entity }) => {
-      await markRelatedGraphChangeNotificationsAsRead(params);
-
-      const response = await updateEntity({
-        variables: {
-          entityId: params.draftEntity.metadata.recordId.entityId,
-          updatedProperties: params.draftEntity.properties,
-          draft: false,
-        },
-      });
-
-      await refetch();
-
-      if (!response.data) {
-        throw new Error("An error occurred accepting the draft entity.");
-      }
-
-      return response.data.updateEntity;
-    },
-    [updateEntity, refetch, markRelatedGraphChangeNotificationsAsRead],
-  );
-
   const value = useMemo<DraftEntitiesContextValue>(
     () => ({
       draftEntities,
@@ -212,17 +106,8 @@ export const DraftEntitiesContextProvider: FunctionComponent<
       refetch: async () => {
         await refetch();
       },
-      discardDraftEntity,
-      acceptDraftEntity,
     }),
-    [
-      draftEntities,
-      draftEntitiesSubgraph,
-      loading,
-      refetch,
-      discardDraftEntity,
-      acceptDraftEntity,
-    ],
+    [draftEntities, draftEntitiesSubgraph, loading, refetch],
   );
 
   return (
